Move sprites, dont use a deck thread
